Divide main page to separate components
