+++ simple-chat/src/chat.html
@@ -56,5 +56,6 @@
                 </button>
             </div>
         </form>
+        <script type="module" src="chat.js"></script>
     </body>
 </html>
