+++ simple-chat/src/scripts/chat-list.js
@@ -1,0 +1,4 @@
+const chatButton = document.querySelector('.main-page__chat');
+chatButton.addEventListener('click', () => {
+    window.location.href = 'http://localhost:8080/chat.html';
+});
