+++ simple-chat/src/scripts/index.js
@@ -1,48 +1,2 @@
 import '../styles/index.css';
-
-const form = document.querySelector('form');
-const input = document.querySelector('.message-input');
-const messageList = document.querySelector('.message-list');
-
-const oldMessages = JSON.parse(localStorage.getItem('messages')) ?? [];
-
-window.addEventListener('load', () => {
-    oldMessages.forEach(displayMessage);
-});
-window.addEventListener('unload', () => saveInLocalStorage());
-form.addEventListener('submit', handleSubmit);
-form.addEventListener('keypress', handleKeyPress);
-
-function handleSubmit(event) {
-    event.preventDefault();
-    const messageText = input.value.trim();
-    if (messageText && messageText.length > 0) {
-        const message = {
-            text: messageText,
-            from: 'Jennifer',
-            time: new Date().toLocaleTimeString('RU', { hour: '2-digit', minute: '2-digit' }),
-        };
-
-        oldMessages.push(message);
-        displayMessage(message);
-        input.value = '';
-    }
-}
-
-function handleKeyPress(event) {
-    if (event.keyCode === 13 && !event.shiftKey) {
-        form.dispatchEvent(new Event('submit'));
-    }
-}
-
-function displayMessage(message) {
-    const messageElement = document.createElement('div');
-    messageElement.classList.add('message');
-    messageElement.innerText = `${message.text}`;
-    messageElement.insertAdjacentHTML('beforeend', `<span class="message-time">${message.from} at ${message.time}</span>`);
-    messageList.appendChild(messageElement);
-}
-
-function saveInLocalStorage(message) {
-    localStorage.setItem('messages', JSON.stringify(oldMessages));
-}
+import './chat-list';
