+++ simple-chat/src/styles/index.css
@@ -1,7 +1,9 @@
 @import url('./message-form.css');
 @import url('./chat-header.css');
 @import url('./messages.css');
-@import url('./main-page.css');
+@import url('./main-page-header.css');
+@import url('./main-page-chats-list.css');
+@import url('./main-page-new-chat-button.css');
 
 html {
     box-sizing: border-box;
