+++ simple-chat/src/styles/main-page-chats-list.css
@@ -1,24 +1,3 @@
-.main-page__header {
-    height: 64px;
-    display: flex;
-    gap: 30px;
-    align-items: center;
-    padding: 12px;
-    background-color: rgb(142 36 170);
-    color: rgb(255 255 255);
-    font-size: 20px;
-}
-
-.main-page__button {
-    background-color: transparent;
-    border: none;
-    color: rgb(255 255 255);
-}
-
-.main-page__title {
-    flex-grow: 1;
-}
-
 .main-page__chats-list {
     height: 100%;
     padding: 12px;
@@ -36,6 +15,7 @@
         "avatar chat-name time"
         "avatar last-message is-readed";
     align-items: center;
+    cursor: pointer;
 }
 
 .main-page__avatar {
@@ -68,21 +48,9 @@
     display: flex;
     justify-content: flex-end;
     grid-area: is-readed;
+    color: rgb(142 36 170);
 }
 
 .check-icon.material-icons {
     font-size: 16px;
-}
-
-.main-page__new-chat-button {
-    border-radius: 50%;
-    width: 50px;
-    height: 50px;
-    border: none;
-    background-color: rgb(255 213 79);
-    color: rgb(117 97 36);
-    cursor: pointer;
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
 }
