+++ simple-chat/src/styles/main-page-header.css
@@ -1,0 +1,20 @@
+.main-page__header {
+    height: 64px;
+    display: flex;
+    gap: 30px;
+    align-items: center;
+    padding: 12px;
+    background-color: rgb(142 36 170);
+    color: rgb(255 255 255);
+    font-size: 20px;
+}
+
+.main-page__button {
+    background-color: transparent;
+    border: none;
+    color: rgb(255 255 255);
+}
+
+.main-page__title {
+    flex-grow: 1;
+}
