+++ simple-chat/src/styles/main-page-new-chat-button.css
@@ -1,0 +1,12 @@
+.main-page__new-chat-button {
+    border-radius: 50%;
+    width: 50px;
+    height: 50px;
+    border: none;
+    background-color: rgb(255 213 79);
+    color: rgb(117 97 36);
+    cursor: pointer;
+    position: fixed;
+    bottom: 20px;
+    right: 20px;
+}
