+++ simple-chat/webpack.config.js
@@ -4,7 +4,6 @@ const path = require('path');
 
 const HTMLWebpackPlugin = require('html-webpack-plugin');
 const MiniCSSExtractPlugin = require('mini-css-extract-plugin');
-const webpack = require('webpack');
 
 const SRC_PATH = path.resolve(__dirname, 'src');
 const BUILD_PATH = path.resolve(__dirname, 'build');
@@ -13,6 +12,7 @@ module.exports = {
     context: SRC_PATH,
     entry: {
         index: './scripts/index.js',
+        chat: './scripts/chat.js'
     },
     output: {
         path: BUILD_PATH,
@@ -62,7 +62,18 @@ module.exports = {
         }),
         new HTMLWebpackPlugin({
             filename: 'index.html',
-            template: './index.html'
-        })
-    ]
+            template: './index.html',
+            chunks: ['index'],
+        }),
+        new HTMLWebpackPlugin({
+            filename: 'chat.html',
+            template: './chat.html',
+            chunks: ['chat'],
+        }),
+    ],
+    devServer: {
+        contentBase: BUILD_PATH,
+        compress: true,
+        port: 8080,
+    }
 };
